refactor: move schemaToType to plugin utils from @hey-api/types
