+++ packages/openapi-ts/src/ir/schema.ts
@@ -1,4 +1,4 @@
-import type { IRSchemaObject } from './ir';
+import type { IRParameterObject, IRSchemaObject } from './ir';
 
 /**
  * Ensure we don't produce redundant types, e.g. string | string.
@@ -63,4 +63,39 @@ export const deduplicateSchema = <T extends IRSchemaObject>({
   }
 
   return schema;
+};
+
+export const irParametersToIrSchema = ({
+  parameters,
+}: {
+  parameters: Record<string, IRParameterObject>;
+}): IRSchemaObject => {
+  const irSchema: IRSchemaObject = {
+    type: 'object',
+  };
+
+  if (parameters) {
+    const properties: Record<string, IRSchemaObject> = {};
+    const required: Array<string> = [];
+
+    for (const name in parameters) {
+      const parameter = parameters[name];
+
+      properties[name] = deduplicateSchema({
+        schema: parameter.schema,
+      });
+
+      if (parameter.required) {
+        required.push(name);
+      }
+    }
+
+    irSchema.properties = properties;
+
+    if (required.length) {
+      irSchema.required = required;
+    }
+  }
+
+  return irSchema;
 };
